refactor: extract project file edit helpers
